Ensure cropped output is truncated
To properly handle the case that a calling app's use of the crop intent
provides an output URI that already exists, e.g. one that is the same
as the input URI, ensure that writes are truncated.

If any apps are affected by this in practice, it's not clear which;
the generic crop intent does not appear to be widely-used, and calls
from AOSP apps or from the system use temporary files for this.

Change-Id: I2a468d9527b679b6aaef6af1521ec0467f58bf60